Create all_types_arrays_performance.cpp
This software builds a huge integer array from the heap, declares the same large array statically (static array), then declares the same large array on the stack (local array variable) (malloc). I fill in the values of these array entries at random. I also used the same array size to ensure that the computation is fair. I called each of the subprograms (functions) a huge number of times (in this example 100,000) and output the amount of time each function took.